analog wheel

+++ src/main/java/com/rkade/WheelDataReport.java
@@ -35,7 +35,7 @@ public final class WheelDataReport extends DataReport {
         if (range == 0) {
             angle = 0;
         } else {
-            angle = value / ((double) Short.MAX_VALUE / (range / 2.0));
+            angle = (double) value / ((double) Short.MAX_VALUE / ((double) range / 2.0));
         }
     }
